# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  kicknroar_backend_internal_dto_request.CreateVenueRequest:
    description: Venue creation request with location, facilities, and contact information
    properties:
      address:
        example: Road 27, Dhanmondi, Dhaka
        type: string
      capacity:
        example: 10
        minimum: 1
        type: integer
      contact_info:
        additionalProperties: true
        type: object
      facilities:
        example:
        - parking
        - changing_room
        - water_dispenser
        items:
          type: string
        type: array
      field_type:
        enum:
        - futsal
        - football
        - astro
        example: futsal
        type: string
      google_place_id:
        example: ChIJN1t_tDeuEmsRUsoyG83frY4
        type: string
      hourly_rate:
        example: 2000
        type: integer
      latitude:
        example: 23.7465
        type: number
      longitude:
        example: 90.376
        type: number
      name:
        example: Dhanmondi Futsal Arena
        maxLength: 100
        minLength: 2
        type: string
      operating_hours:
        additionalProperties: true
        type: object
      surface_type:
        enum:
        - grass
        - artificial
        - concrete
        example: artificial
        type: string
    required:
    - address
    - capacity
    - field_type
    - latitude
    - longitude
    - name
    type: object
...
  kicknroar_backend_internal_dto_request.UpdateVenueRequest:
    description: Venue update request - all fields are optional
    properties:
      address:
        example: Road 27, Dhanmondi, Dhaka - Updated
        type: string
      capacity:
        example: 12
        minimum: 1
        type: integer
      contact_info:
        additionalProperties: true
        type: object
      facilities:
        example:
        - parking
        - changing_room
        - water_dispenser
        - shower
        items:
          type: string
        type: array
      field_type:
        enum:
        - futsal
        - football
        - astro
        example: futsal
        type: string
      hourly_rate:
        example: 2500
        type: integer
      latitude:
        example: 23.7465
        type: number
      longitude:
        example: 90.376
        type: number
      name:
        example: Dhanmondi Futsal Arena - Updated
        maxLength: 100
        minLength: 2
        type: string
      operating_hours:
        additionalProperties: true
        type: object
      surface_type:
        enum:
        - grass
        - artificial
        - concrete
        example: artificial
        type: string
    type: object
...
  kicknroar_backend_internal_dto_response.ErrorDetail:
    description: Detailed error information
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        example:
        - email is required
        - password must be at least 8 characters
        items:
          type: string
        type: array
      message:
        example: Invalid request data
        type: string
    type: object
  kicknroar_backend_internal_dto_response.ErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorDetail'
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@kickandroar.com
    name: API Support
  description: Football match coordination platform for Dhaka, Bangladesh
  license:
    name: Proprietary
  title: Kick&Roar API
  version: "1.0"
paths:
...
  /api/v1/venues:
    get:
      consumes:
      - application/json
      description: Get a list of all venues with pagination
      parameters:
      - description: 'Number of results per page (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: List venues
      tags:
      - venues
    post:
      consumes:
      - application/json
      description: Create a new football venue with location, facilities, and other
        details
      parameters:
      - description: Venue creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kicknroar_backend_internal_dto_request.CreateVenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new venue
      tags:
      - venues
  /api/v1/venues/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get venue by ID
      tags:
      - venues
    put:
      consumes:
      - application/json
      description: Update venue details (only owner can update)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Venue update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kicknroar_backend_internal_dto_request.UpdateVenueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update venue
      tags:
      - venues
  /api/v1/venues/{id}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a venue (only owner can upload)
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file (max 5MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Upload venue image
      tags:
      - venues
  /api/v1/venues/nearby:
    get:
      consumes:
      - application/json
      description: Find venues within a specified radius from given coordinates
      parameters:
      - description: Latitude
        format: float64
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        format: float64
        in: query
        name: lng
        required: true
        type: number
      - description: 'Search radius in kilometers (default: 5)'
        format: float64
        in: query
        name: radius
        type: number
      - description: Filter by field type (futsal, football, astro)
        in: query
        name: field_type
        type: string
      - description: 'Maximum number of results (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find nearby venues
      tags:
      - venues
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"